backend/voiceloop-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
openai==1.12.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
whisper-openai==0.1.0
chromadb==0.4.22
sentence-transformers==2.5.1
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
accelerate==0.24.1
